cookidoo_today/app/requirements.txt
fastapi>=0.110
uvicorn>=0.27
aiohttp>=3.9
Pillow>=10.0
cookidoo-api==0.15.0